cloud_backend/requirements.txt
aiohttp==3.11.18
google-cloud-firestore==2.20.1
google-cloud-logging==3.12.1
